# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Compscore application",
        "title": "Compscore API",
        "contact": {
            "name": "1nv8rzim",
            "url": "https://github.com/compsore/compscore/issues"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/status/round/{round}": {
            "get": {
                "description": "Get all status entries for a given round",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Get all status entries for a given round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Round number",
                        "name": "round",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Status"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/status/team/{team}/round/{round}": {
            "get": {
                "description": "Get all status entries for a given team and round",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "status"
                ],
                "summary": "Get all status entries for a given team and round",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Team ID",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Round number",
                        "name": "round",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Status"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Error": {
            "description": "response of login request",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.Status": {
            "description": "status of a check",
            "type": "object",
            "properties": {
                "edges": {
                    "type": "object",
                    "properties": {
                        "check": {},
                        "round": {},
                        "team": {}
                    }
                },
                "error": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ServiceAuth": {
            "description": "JWT for authentication",
            "type": "apiKey",
            "name": "auth",
            "in": "cookie"
        }
    }
}